pyproject.toml
[project]
name = "doc2data"
version = "0.1.0"
description = "doc2data using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.186.1",
    "pdfplumber>=0.11",
    "psycopg2-binary>=2.9",
    "python-dateutil",
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
    "humanize>=4.12.3",
    "pandas>=2.3.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.0",
    "streamlit>=1.45.1",
]

[project.scripts]
kickoff = "doc2data.main:kickoff"
run_crew = "doc2data.main:kickoff"
plot = "doc2data.main:plot"
run_with_trigger = "doc2data.main:run_with_trigger"
run_local = "doc2data.main:run_local"
run_gdrive = "doc2data.main:run_gdrive"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
